MC-Streamer/requirements.txt
xlib
pillow
websockets
opencv-python
ffmpeg-python
ws4py
